Refactor CORS configuration and cleanup unnecessary code
The CorsConfig class was completely removed, and its functionality was transferred directly into the SecurityConfig class in order to consolidate configuration for easier maintenance and avoid redundancy. The CORS configuration is now simplified to accept all origins which will facilitate request handling and makes it more docker-friendly. The AuthService.java was cleaned up by removing unnecessary blank lines for a better-looking code.